Generate cosmology animation

+++ chop_data.py
@@ -1,16 +1,18 @@
 from __future__ import annotations
-from typing import Iterable, Any
-import shutil
-from tqdm import tqdm
-import os
-import stat
+
 import datetime
 import itertools
+import os
 import re
+import shutil
+import stat
+from pathlib import Path
+from typing import Any, Iterable
+
 import dask.bag  # type: ignore
 import dask.diagnostics  # type: ignore
-from pathlib import Path
 import yt  # type: ignore
+from tqdm import tqdm
 
 
 def consume(x: Iterable[Any]) -> None:
@@ -28,21 +30,35 @@ def mtime(path: Path) -> datetime.datetime:
 def plot_cosmology(data_dir: Path, output_dir: Path) -> None:
     import matplotlib.animation  # type: ignore
 
-    ts = yt.load(data_dir / "DD????/data????")
-    plot = (
-        yt.ProjectionPlot(ts[0], axis="z", fields=dm_field, method="integrate",)
-        .annotate_timestamp(corner="upper_left", redshift=True, draw_inset_box=True)
-        .annotate_scale(corner="upper_right")
-        .set_log(dm_field, True, symlog_auto=True)
-    )
-    animation = matplotlib.animation.FuncAnimation(
-        plot.plots[dm_field].figure,
-        lambda i: plot._switch_ds(ts[i]),
-        frames=len(ts),
-        fps=1,
-    )
-    animation.save("projection.gif")
-    animation.save("projection_%04d.png")
+    try:
+        ts = yt.load(str(data_dir / "DD????/data????"))
+    except FileNotFoundError:
+        print("None found for", str(data_dir / "DD????/data????"))
+        return
+    if len(ts) > 0:
+        for i, ds in enumerate(ts):
+            plot = (
+                yt.ProjectionPlot(ds, axis="z", fields=dm_field, method="integrate",)
+                .annotate_timestamp(corner="upper_left", redshift=True, draw_inset_box=True)
+                .annotate_scale(corner="upper_right")
+                .set_log(dm_field, True, symlog_auto=True)
+                .save(output_dir / f"projection_{i:04d}.png")
+            )
+        # plot = (
+        #     yt.ProjectionPlot(ts[0], axis="z", fields=dm_field, method="integrate",)
+        #     .annotate_timestamp(corner="upper_left", redshift=True, draw_inset_box=True)
+        #     .annotate_scale(corner="upper_right")
+        #     .set_log(dm_field, True, symlog_auto=True)
+        # )
+        # animation = matplotlib.animation.FuncAnimation(
+        #     plot.plots[dm_field].figure,
+        #     lambda i: plot._switch_ds(ts[3 * i]),
+        #     frames=len(ts) // 3,
+        #     interval=1000,
+        # )
+        # animation.save("projection.gif")
+        # print("hi", len(ts))
+        # animation.save("projection_%04d.png")
 
 
 def chop_frame(
@@ -81,14 +97,14 @@ def chop_nn_class_dir(
     cosmology_dir = nn_class_data_dir / "cosmology"
     chopped_dir = nn_class_data_dir / "chopped"
 
-    if mtime(raw_dir) > mtime(cosmology_dir):
+    if cosmology_dir.exists() and mtime(raw_dir) > mtime(cosmology_dir):
         shutil.rmtree(cosmology_dir)
     if not cosmology_dir.exists():
         cosmology_dir.mkdir()
-        print("going on cosmology")
+        print("going on plotting")
         plot_cosmology(raw_dir, cosmology_dir)
 
-    if mtime(raw_dir) > mtime(chopped_dir):
+    if chopped_dir.exists() and  mtime(raw_dir) > mtime(chopped_dir):
         shutil.rmtree(chopped_dir)
     if not chopped_dir.exists():
         print("going on chopping")
@@ -107,6 +123,10 @@ if __name__ == "__main__":
     yt.set_log_level("warning")
     print(f"main({nn_data_dir}, {voxels_per_side}, {padding})")
     with dask.diagnostics.ProgressBar():
-        dask.bag.from_sequence(nn_data_dir.iterdir()).map(
+        paths = [
+            nn_data_dir / "train/high",
+        ]
+        
+        dask.bag.from_sequence(paths).map(
             chop_nn_class_dir, voxels_per_side=voxels_per_side, padding=padding
-        )
+        ).compute()
